+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-### **(V.1.2.0 Changes) (1.17.0 Minecraft)**
+### **(V.1.3.0 Changes) (1.17.1 Minecraft)**
+
+##### Misc:
+Added config to let you specify what structure should have pieces should be raised up separately from other pieces in the same structure.
+  Think, villages.
+
+
+### **(V.1.2.1 Changes) (1.17.1 Minecraft)**
+
+##### Misc:
+Tried fixing a crash.
+
+
+### **(V.1.2.0 Changes) (1.17.1 Minecraft)**
 
 ##### Pillars:
 Fixed pillar code to be more accurate and not be messed up by caves and ravines.
@@ -9,13 +22,13 @@ Split removeWorldBottomStructures into removeStructuresOffIslands and removeWorl
  removeStructuresOffIslands now will do better (hopefully) to remove any structure that goes off the island.
 
 
-### **(V.1.1.0 Changes) (1.17.0 Minecraft)**
+### **(V.1.1.0 Changes) (1.17.1 Minecraft)**
 
 ##### Misc:
 More options/configs and allow proper floating villages and a blacklist etc.
 
 
-### **(V.1.0.0 Changes) (1.17.0 Minecraft)**
+### **(V.1.0.0 Changes) (1.17.1 Minecraft)**
 
 ##### Misc:
 First Creation